solucao parcial de validacoes e bloqueios de agencia 1, correntista 1 e contas do correntista 1

+++ src/main/java/com/banco/bluebank/service/ContaService.java
@@ -9,6 +9,7 @@ import com.banco.bluebank.repository.AgenciaRepository;
 import com.banco.bluebank.repository.ContaRepository;
 import com.banco.bluebank.repository.CorrentistaRepository;
 import com.banco.bluebank.security.SysSecurity;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -21,6 +22,8 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.validation.ConstraintViolationException;
 import java.time.OffsetDateTime;
 import java.time.format.DateTimeParseException;
+
+import static com.banco.bluebank.service.CorrentistaService.CORRENTISTA_BLUEBANK;
 
 @Service
 public class ContaService {
@@ -83,6 +86,10 @@ public class ContaService {
         Correntista correntista = correntistaRepository.findById(conta.getCorrentista().getId())
                 .orElseThrow(() -> new CorrentistaNaoEncontradoException(finalConta.getCorrentista().getId()));
 
+        if(correntista.getId() == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Id de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
+
         conta.setIdCorrentista(correntista.getId());
 
         Conta finalConta1 = conta;
@@ -102,19 +109,29 @@ public class ContaService {
     }
 
     @Transactional(readOnly = false)
-    public Conta atualizar(Long numeroConta, Conta contaAtual) {
+    public Conta atualizar(Long numeroConta, Conta conta) {
 
-        Conta conta = buscar(numeroConta);
+        Conta contaAtual = buscar(numeroConta);
+        Long idCorrentistaArmazenado = contaAtual.getCorrentista().getId();
+
+        BeanUtils.copyProperties(conta, contaAtual, "numeroConta", "dataCadastro");
+
+        contaAtual.setIdCorrentista(contaAtual.getCorrentista().getId());
+        contaAtual.setIdAgencia(contaAtual.getAgencia().getId());
 
         Correntista correntista = correntistaRepository.findById(contaAtual.getIdCorrentista())
                 .orElseThrow(() -> new CorrentistaNaoEncontradoException(contaAtual.getIdCorrentista()));
+
+        if(correntista.getId() == CORRENTISTA_BLUEBANK || idCorrentistaArmazenado == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de conta bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
 
         Agencia agencia = agenciaRepository.findById(contaAtual.getIdAgencia())
                 .orElseThrow(() -> new AgenciaNaoEncontradaException(contaAtual.getIdAgencia()));
 
         conta.setSenha(passwordEncoder.encode(conta.getSenha()));
 
-        conta = contaRepository.save(conta);
+        conta = contaRepository.save(contaAtual);
         conta.setCorrentista(correntista);
         conta.setAgencia(agencia);
         return conta;
@@ -123,7 +140,18 @@ public class ContaService {
 
     public void excluir(Long numeroConta) {
         if(numeroConta==null) throw new DadoRequeridoException("O número da conta é requerido");
+
+        Conta conta = buscar(numeroConta);
+
+        Correntista correntista = correntistaRepository.findById(conta.getIdCorrentista())
+                .orElseThrow(() -> new CorrentistaNaoEncontradoException(conta.getIdCorrentista()));
+
+        if(correntista.getId() == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de conta bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
+
         Long numeroContaSemDigito = contaUtils.verificaNumeroConta(numeroConta);
+
         try {
             contaRepository.deleteById(numeroContaSemDigito);
