+++ src/main/java/com/banco/bluebank/service/CorrentistaService.java
@@ -1,9 +1,6 @@
 package com.banco.bluebank.service;
 
-import com.banco.bluebank.exceptionhandler.exceptions.ContatoNaoEncontradoException;
-import com.banco.bluebank.exceptionhandler.exceptions.CorrentistaNaoEncontradoException;
-import com.banco.bluebank.exceptionhandler.exceptions.EnderecoNaoEncontradoException;
-import com.banco.bluebank.exceptionhandler.exceptions.EntidadeEmUsoException;
+import com.banco.bluebank.exceptionhandler.exceptions.*;
 import com.banco.bluebank.model.ContatoCliente;
 import com.banco.bluebank.model.Correntista;
 import com.banco.bluebank.model.Endereco;
@@ -20,11 +17,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Locale;
+import java.util.Objects;
 
 @Service
 public class CorrentistaService {
 
     private static final String MSG_CORRENTISTA_EM_USO = "Correntista de id %d não pode ser removido, pois está em uso";
+    public static final long CORRENTISTA_BLUEBANK = 1L;
 
     @Autowired
     private CorrentistaRepository correntistaRepository;
@@ -38,18 +37,18 @@ public class CorrentistaService {
     @Transactional(readOnly = false)
     public Correntista create(Correntista correntista) {
 
-        if(correntista.getTipoPessoa()!=null) correntista.setTipoPessoa(correntista.getTipoPessoa().toUpperCase());
+        if (correntista.getTipoPessoa() != null) correntista.setTipoPessoa(correntista.getTipoPessoa().toUpperCase());
 
         Correntista novoCorrentista = correntistaRepository.save(correntista);
 
         List<Endereco> enderecos = correntista.getEnderecos();
-        for(Endereco endereco: enderecos) {
+        for (Endereco endereco : enderecos) {
             endereco.setIdCorrentista(novoCorrentista.getId());
             endereco = enderecoRepository.save(endereco);
         }
 
         List<ContatoCliente> contatos = correntista.getContatos();
-        for(ContatoCliente contato: contatos) {
+        for (ContatoCliente contato : contatos) {
             contato.setIdCorrentista(novoCorrentista.getId());
             contato = contatoClienteRepository.save(contato);
         }
@@ -60,7 +59,10 @@ public class CorrentistaService {
     @Transactional(readOnly = false)
     public Correntista update(Correntista correntista) {
 
-        if(correntista.getTipoPessoa()!=null) correntista.setTipoPessoa(correntista.getTipoPessoa().toUpperCase());
+        if (correntista.getId() == CORRENTISTA_BLUEBANK)
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+
+        if (correntista.getTipoPessoa() != null) correntista.setTipoPessoa(correntista.getTipoPessoa().toUpperCase());
 
         return correntistaRepository.save(correntista);
 
@@ -74,10 +76,15 @@ public class CorrentistaService {
 
     public Correntista buscar(Long correntistaId) {
         return correntistaRepository.findById(correntistaId)
-                .orElseThrow( () -> new CorrentistaNaoEncontradoException(correntistaId));
+                .orElseThrow(() -> new CorrentistaNaoEncontradoException(correntistaId));
     }
 
+    @Transactional(readOnly = false)
     public void excluir(Long correntistaId) {
+
+        if (correntistaId == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
 
         buscar(correntistaId);
 
@@ -106,6 +113,10 @@ public class CorrentistaService {
     @Transactional(readOnly = false)
     public List<Endereco> adicionarEndereco(Long correntistaId, Endereco endereco) {
 
+        if (correntistaId == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
+
         Correntista correntista = this.buscar(correntistaId);
 
         endereco.setIdCorrentista(correntistaId);
@@ -117,6 +128,10 @@ public class CorrentistaService {
 
     @Transactional(readOnly = false)
     public List<ContatoCliente> adicionarContato(Long correntistaId, ContatoCliente contato) {
+
+        if (correntistaId == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
 
         Correntista correntista = this.buscar(correntistaId);
 
@@ -130,7 +145,19 @@ public class CorrentistaService {
     @Transactional(readOnly = false)
     public void excluirEndereco(Long correntistaId, Long enderecoId) {
 
+        if (correntistaId == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
+
         buscar(correntistaId);
+
+        Endereco endereco = enderecoRepository.findById(enderecoId)
+                .orElseThrow(() -> new EnderecoNaoEncontradoException(enderecoId));
+
+        if (!Objects.equals(endereco.getCorrentista().getId(), correntistaId)) {
+            throw new EnderecoNaoEncontradoException(
+                    String.format("O endereco especificado %d nao corresponde ao correntista %d", enderecoId, correntistaId));
+        }
 
         try {
             enderecoRepository.deleteById(enderecoId);
@@ -144,7 +171,19 @@ public class CorrentistaService {
     @Transactional(readOnly = false)
     public void excluirContato(Long correntistaId, Long contatoId) {
 
+        if (correntistaId == CORRENTISTA_BLUEBANK) {
+            throw new RecursoComBloqueioException("Recurso de correntista bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+        }
+
         buscar(correntistaId);
+
+        ContatoCliente contato = contatoClienteRepository.findById(contatoId)
+                .orElseThrow(() -> new ContaNaoEncontradaException(contatoId));
+
+        if (!Objects.equals(contato.getCorrentista().getId(), correntistaId)) {
+            throw new ContatoNaoEncontradoException(
+                    String.format("O contato especificado %d nao corresponde ao correntista %d", contatoId, correntistaId));
+        }
 
         try {
             contatoClienteRepository.deleteById(contatoId);
